requirements.txt
# Core ML
torch>=2.0.0
numpy>=1.24.0

# Chess
python-chess>=1.9.0

# Data pipeline
pandas>=2.0.0
pyarrow>=12.0.0
h5py>=3.8.0
zstandard>=0.21.0
requests>=2.31.0
tqdm>=4.65.0

# Training
tensorboard>=2.13.0

# Evaluation
matplotlib>=3.7.0
seaborn>=0.12.0

# Web backend
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Web frontend (managed by npm)
# react, react-chessboard, chess.js, recharts

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0

# ONNX export
onnx>=1.14.0
onnxruntime>=1.15.0

# HuggingFace datasets
datasets>=2.14.0
